Allow API Calls Without User Agents
Current implementation does not handle sessions missing the user agent header. It is not uncommon for requests made by API's to lack a user agent in their headers. According to RFC 7231, the user agent header is not mandatory. 

This line will produce an error resulting in a 500 response from the server if there does not exist a user agent for the session and user agent restriction is enabled. In order to avoid the error, use the .get method which will substitute None if a key cannot be found. This fix prevents the error while continuing to allow user agent restriction for sessions based on browser and appropriately handle an API-based session if it lacks a user agent.